micromegas_transit

Module uuid_utils

uuid encoding

Functions

opt_uuid_from_string

Deserializes an optional UUID from a string using serde, returning None for empty strings

opt_uuid_to_string

Serializes an optional UUID to a string using serde, using empty string for None

parse_optional_uuid

Parse an optional UUID from a string, returning None if the string is empty

uuid_from_string

Deserializes a UUID from a string using serde

uuid_to_string

Serializes a UUID to a string using serde