Crate micromegas_transit

transit library provides fast binary serialization for Plain Old Data structures

Modules

prelude

string_codec

string encoding

uuid_utils

uuid encoding

value

json-like variant

Structs

DynString

LegacyDynString

Member

QueueIterator

StaticStringDependency

StaticStringDependency serializes the value of the pointer and the contents of the string It is designed to be wire-compatible with the unreal instrumentation

UserDefinedType

Utf8StaticStringDependency

Utf8StaticStringDependency serializes the value of the pointer and the contents of the string It should not be part of the event - it’s the dependency of the StringId

Enums

InProcSize

Helps speed up the serialization of types which size is known at compile time.

Traits

HeterogeneousQueue

InProcSerialize

Reflect

Functions

advance_window

parse_object_buffer

parse_pod_instance

read_advance_string

Parse string from buffer, move buffer pointer forward.

read_any⚠

Helper function to read a u* pointer to a value of type T.

read_consume_pod

read_dependencies

write_any

Type Aliases

CustomReader

CustomReaderMap