micromegas_transit

Trait InProcSerialize

pub trait InProcSerialize: Sized {
    const IN_PROC_SIZE: InProcSize = _;

    // Provided methods
    fn get_value_size(&self) -> Option<u32> { ... }
    fn write_value(&self, buffer: &mut Vec<u8>) { ... }
    unsafe fn read_value(window: &[u8]) -> Self { ... }
}

Provided Associated Constants

const IN_PROC_SIZE: InProcSize = _

Provided Methods

fn get_value_size(&self) -> Option<u32>

fn write_value(&self, buffer: &mut Vec<u8>)

unsafe fn read_value(window: &[u8]) -> Self

Safety

This is called from the serializer context that that uses value_size call to make sure that the proper size is used

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.